requirements.txt
setuptools==80.9.0
Flask==3.0.0
Flask-RESTful==0.3.10
Werkzeug==3.0.1
rasterio==1.4.3
gunicorn==21.2.0
pytest==7.4.3
pillow==11.3.0